Update surface-hub-teams-rooms.md
Line 15: , which were > that (@CoveMiner - Do you approve this editorial change? I made it for concision and to avoid the invisible "Microsoft released." Now it released itself, but doesn't need a tetchy dependant clause and the companion comma.)

Lines 33, 35, 45 and 73: Revised alt-text to comply with the criteria.

Line 66, 77-78: Align image with the associated step.

Line 80: Add space after the step number and change the step number to 1. Check that indenting the note enabled the autonumbering for the subsequent item.